restructing the auth modules
